final, calculator works with division floats, redirects back to calculator

+++ calculator.rb
@@ -1,18 +1,19 @@
 def calculator(calc, num1, num2)
     if  calc == "plus"
-        answer = num1.to_i + num2.to_i
+        answer = num1.to_f + num2.to_f
     end
     
     if calc == "minus"
-        answer = num1.to_i - num2.to_i
+        answer = num1.to_f - num2.to_f
     end
 
     if calc == "times"
-        answer = num1.to_i * num2.to_i
+        answer = num1.to_f * num2.to_f
     end
 
     if calc == "divided by"
-        answer = num1.to_i / num2.to_i
+        answer = num1.to_f / num2.to_f
     end
+
     answer.to_s
 end
